Adjusted Zoom controls
Added a zoom_factor to apply in each case.

ZOOM_MAX & ZOOM_MIN now defined in fdf.h

Ideally these values would be dynamic in response to the map size but
would have to be included in the data structure to allow this whilst
staying within the confines of the subject norm.